feat: created IRGenerator interface
